Make LazilySavebalePath an abstract base class
The class already relied on methods implemented in its subclasses, so
implementing it as an ABC is a natural choice. The change will prevent
it from being accidentally instantiated and, with the help of linters
(such as pylama), it will provide better error handling when writing
subclasses of the abstract class.