Fix newly exposed test failures
This commit fixes test failures exposed by the previous commit.

It turns out there were several errors in the content of the expected
string used in the assertion about the error message. They were masked
by the placement of the assertion call - I had put it in assertRaises
context manager block, right after the call that was expected to raise
ConfigKeyError, and I didn't realize this would make the call
unreachable if the exception was raised.